clk: qcom: smd-rpm: HACK! Keep CXO up for MSM8998 as death WA
This platform dies if we shut down the CXO (the board stays
alive, but peripherals will not work, such as display and some
i2c devices).

This is a temporary workaround to disallow RPM XO shutdown
which will consequently disable RPM VMIN and other ultra low
power modes during sleep.

Battery consumption will be higher.
This is the only way to get the SoMC Yoshino platform to work
while a solution gets found (which will take a lot of time...).