arm64: dts: msm8998: Move energy-costs to v2 and update definitions
Follow the new format for 4.14 by which we define the busy cost against the
CPU frequency. In addition to this, we move the definition of the data to v2
and properly have sched-energy acknowledge the data. The capacity-dmips-mhz
values are also added based upon the CPUs' highest capacity value which we
now replace with frequencies.

Let's also fix some indentions in msm8998.dtsi.

Signed-off-by: ederekun <sedrickvince@gmail.com>