msm: mdss: HACK: add function to shutdown the framebuffer [1/3]
This is the first of three commits that will implement a mechanism to shutdown
the primary framebuffer, leading Android to resetting it and therefore
resetting the MDP clocks, bypassing a display initialization issue.

In this commit, we add the 'fb_blank_powerdown' function to the MDSS framebuffer
driver: this method can be called by any driver importing the fb.h header.

We will use it to shutdown the framebuffer as soon as the Android partitions
are mounted, so that Android is ready to start the boot animation.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>