arm64: DT: msm8998: Repair thermal zones and remove LMH lite
Sony did not update the DT nodes for pm8998_adc_tm, hence our ADC
thermal nodes not working. This commit fixes all of our currently
defined nodes accordingly based on documentation and newer SoCs.
In addition to this, remove LMH lite as it is likely only useful
with HW platform mitigation which we no longer depend on.

Signed-off-by: ederekun <sedrickvince@gmail.com>