BACKPORT: sched/fair: Make EAS wakeup placement consider uclamp restrictions
This is a modified version of:

https://github.com/Juansheng/kernel_oneplus_sm8250/commit/ecce1cf84a2e719fc88267123a6ce689594a530e#diff-75a417d0e6ebb27175a48c6810294ab6e21847776b1e46c62de9cccfb428883eR7018

The best equivalent code path that reflect select_cpu_candidates is
k4.14 find_best_target, wherein the same utilization calculation is
performed along with additional adjustments due to WALT etc. This
is as close as we can get to the actual patch without completely
backporting the entirety of the fair scheduler from k4.19.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>