PM: sleep: introduce suspend_deny_list [HACK]
It was observed that, apparently randomly, the MSM8998 devices enter deep
sleep and they cannot be wake up again.

A device in this state is not completely dead (notification LED continues
to work if it was working at the moment the problem appears, for example),
but no input can wake up the device (e.g. power button, fingerprint
reader, a phone call).

This may happen once a day, or once a week.
At the moment, the reason of this behaviour is not clear.

The only way to recover the device is to reboot it manually keeping the
power button pressed, but after this the pstore is empty, and attempts
to gather logs (e.g. writing it to persisted storage) only provided
logs ending with the line 'PM: PM: suspend entry' line.

It was find out that disabling the suspend for all devices apparently 'fixes'
the problem (the issue was not observed anymore - a month - on devices which
were showing the issue once a day before the patch).

Disabling the suspend for all devices was causing higher power consumption
in this new 'deep sleep' state, so we started to 'bisect' the devices entering
suspend: the initial list was ~1900 devices long, and we now reduced it to
35 devices (after 5 months of test).

This commit adds a list to allow us to exclude devices that we don't want
to enter suspend: this will prevent the kernel to call the suspend/resume
callbacks of the drivers managing these devices.

No increase in power consumption was observed now compared to a device entering
'full deep sleep'.

Ideally, we should continue to bisect until we find the driver causing the problem.

The size of the current list is printed once during boot.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>