[workaround] power: smb-lib: make sure input current limit is correct after the switch from dash to normal charge
It has been observed that the charging process stops 'randomly' when using
a dash charger.

After some investigation, it turned out that this is not really random:
whenever the device switches from dash charge to normal charge (for whatever
reason, for example it reaches a high charge level - 70%) the icl
(input current limit) may be set to 100mA.

This is visible from the healthd log:
healthd: battery l=83 v=4183 t=28.3 h=2 st=4 c=-82000 fc=2819000 cc=1 chg=a

(82mA here).

Printing the current limit in this situation reports '100000' which is 100mA.

We implement here a simple workaround which has been effective during testing:
whenever we have switched from dash to normal charge, we verify the current limit
and if it is too low, we increase it to 1.5A: this is the current limit
whenever the switch happens successfully (this has been verified printing
the limit after a successful switch).

The real cause of the issue is not clear at the moment.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>