arm64: DT: msm8998: Fix removed_region2 size
The initial calculation by Kholk was incorrect which had him put out
another patch that increased removed_region1. It follows that the
calculation of removed_region2 as well was more than necessary as
he made the same wrong assumption as he did with removed_region1.

The end of the removed_region in downstream from 0x85800000 is to
0x88F00000. The current configuration sets it to 0x89500000.
Correct it.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>