clk: qcom: gcc-msm8998: Avoid turning off hmss_ahb_clk during certain LPM states
The clock driver sets the sleep_ena bit to allow the hmss_ahb_clk
to be disabled by hardware during certain low power modes. The
PCIe controller however might need to access some registers that
need this hmss_ahb_clk to be on. Remove the additional settings
in the clock driver to resolve the issue.

CRs-Fixed: 994609
Change-Id: Ib486a27f2e1c2d2231f8bedcb4ee8b39381cbd25
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
[Apply to msm8998]
Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>