sched/fair: Fix idle CPU selection util_fits_cpu logic
Prior to this patch, we evaluate whether the new usage fits the
CPU's capacity. But with the addition of UCLAMP, we need to consider
the clamps of the CPU and the task with util_fits_cpu. This required
us to replace the logic comparing the usage towards the capacity.

How this was resolved is greatly flawed. We did not remove the entire
capacity comparison logic which should be done entirely by util_fits_cpu.
We should actually use util_fits_cpu to decide on whether we should go
towards the next sched_group. And that if the new_usage does fit the CPU,
we should consider it as the best_idle_cpu if it has sufficient idle_idx.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>