clk: qcom: gpucc-msm8998: Complete and cleanup clocks configuration
Let's complete all the missing clock configurations for GPU and cleanup
to ensure clear and proper configuration.

Primarily, gpucc_cxo_clk (gpucc_xo) has been renamed and configured
completely as it was missing its enable_reg and enable_mask which can
cause issues when the hardware expects a vote on the branch. Also follow
newer QCOM SoCs configuration with the clk framework wherein we decouple the
gpucc_cxo_clk from the other clocks as we only need to vote on this
clock, use bi_tcxo_ao as a parent to the GPU clocks which require clock
rate scaling.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>