arm64: DT: msm-arm-smmu-8998: Disable ATOS for SMMU-v2 IOMMUs
The arm-smmu driver in SDM845 disables ATOS in QSMMUV2 model of which
was later then converted to a DT property. It was stated that the model
does not support ATOS debugging in tandem with other features such as
power collapse according to documentation. Follow suit and avoid
possible issues caused by this feature.

Reference:
https://github.com/LineageOS/android_kernel_xiaomi_sdm845/commit/62ba192899ed8f5665fa68aa3d0adb30c8a2bb5c#diff-a80b60637c0a7753a84657e546fac56eb7df91d8b351735786aa98030cff32de

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>